Updated to recognise GEDCOM 5.5.1 Draft standard.
Modifications in Version 5.5.1 (From the Draft standard PDF)
!   The following tags were added:
    EMAIL   electronic mailing address
    FAX FAX address
    FACT    A fact or characteristic.
    FONE    Phonetic variation of a text.
    ROMN    Romanized variation of a text.
    WWW Web Home page address.
    MAP Pertaining to maps.
    LATI    value of a latitudinal coordinate pertaining to the place
of an event
    LONG    value of a longitudinal coordinate pertaining to the place
of an event.
!   The following tag was removed: BLOB (but not from the parsers, as
we want to recognize the 5.5 standard)
In detail
!   Added continuation tags to the header copyright tags (see
<<HEADER>>, page 23.)
!   Added email, fax, and web page addresses to the address structure
(see <<ADDRESS_STRUCTURE>>, page 31.)
!   Added a status tag to the child to family link (see
<<CHILD_TO_FAMILY_LINK>>, page 31.)
!   Added a restriction notice tag to the family record to allow a
source database to indicate why data may not have been supplied in the
transmission. (see <<FAMILY_RECORD>>, page 24.)
!   Also added a restriction notice tag to the <<EVENT_DETAIL>>
structure page 32 to allow an event to be flagged so that it can be
treated in special ways such as not to be printed on reports.
!   Added additional subordinate structure to the personal name
structure (see <<PERSONAL_NAME_STRUCTURE>>, page 38.) These changes are
in preparation for handling more varied cultures as we move into the
unicode character set environment.
!   Added subordinate map coordinates and other additional changes to
the place structure (see <<PLACE_STRUCTURE>>, page 38.) These changes
are in preparation for handling more varied cultures as we move into
the unicode character set environment and to allow recording of map
coordinates to places such as burial cites.
!   Added a subordinate affiliated religion tag to the event detail
substructure (see <<EVENT_DETAIL>>, page 32.)
!   Added a generic FACT tag to the individual attribute structure.
Previously, the generic EVEN tag was used. The FACT was added to give a
semantic difference between generic events and generic facts or
characteristics (see <<INDIVIDUAL_ATTRIBUTE_STRUCTURE>>, page 33.)
!   Removed the option for encoding embedded multimedia objects. A file
reference to a multimedia file and its subordinate format and media
types were added to the multimedia record. Multiple file references can
now be used to group related multimedia objects. This changed the
multimedia link by placing the FORM tag subordinate to the FILE tag
rather than at the same level. The BLOB tag was eliminated. See FILE
tag and its subordinate FORM tag used in the <<MULTIMEDIA_RECORD>> page
26 and the <<MULTIMEDIA_LINK>> page 37.